+++ sdks/wasm/debug.html
@@ -20,7 +20,7 @@
 			},
 
 			init: function () {
-				this.int_add = Module.mono_bind_static_method ("[sample] DebuggerTests.ValueTypesTest:MethodWithLocalStructs");
+				this.int_add = Module.mono_bind_static_method ("[sample] Math:IntAdd");
 				this.output = document.getElementById ("output");
 				this.button = document.getElementById ("button");
